Move ForksSchedule to common package
